backend/pyproject.toml
[project]
name = "fastemplate"
version = "0.1.0.dev0"
description = "Template project for FastAPI applications"
requires-python = ">=3.11,<4.0"
license = { text = "Apache-2.0" }
authors = [{ name = "François-Guillaume Fernandez", email = "fg-feedback@protonmail.com" }]
maintainers = [{ name = "François-Guillaume Fernandez", email = "fg-feedback@protonmail.com" }]
readme = "README.md"
repository = { url = "https://github.com/frgfm/fastemplate" }
keywords = ["backend", "api", "template", "fastapi"]
dependencies = [
    "fastapi>=0.109.1,<1.0.0",
    "sqlmodel>=0.0.23,<0.1.0",
    "pydantic-settings>=2.0.0,<3.0.0",
    "requests>=2.32.0,<3.0.0",
    "bcrypt>=4.2.0,<5.0.0",
    "uvicorn>=0.23.0,<1.0.0",
    "asyncpg>=0.29.0,<1.0.0",
    "PyJWT[crypto]>=2.8.0,<3.0.0",
    "python-multipart>=0.0.7,<1.0.0",
    # cf. https://github.com/boto/boto3/issues/4435
    "boto3>=1.26.0,<1.36.0",
    "email-validator>=2.0.0,<3.0.0",
    "alembic>=1.15.2,<2.0.0",
    # Optional
    "sentry-sdk[fastapi]>=2.14.0,<3.0.0",
    "logfire[fastapi,system-metrics]>=3.0.0,<4.0.0",
    "posthog>=3.0.0,<4.0.0"
]

[project.optional-dependencies]
test = [
    "pytest>=8.3.3,<9.0.0",
    "pytest-asyncio>=0.17.0,<1.0.0",
    # https://github.com/geekan/MetaGPT/issues/1617
    "httpx>=0.23.0,<0.28.0",
    "pytest-cov>=4.0.0,<5.0.0",
    "pytest-pretty>=1.0.0,<2.0.0",
    "requests>=2.32.0,<3.0.0",
    "pillow>=9.4.0"
]

[tool.uv]
dev-dependencies = [
    "ruff==0.11.9",
    "mypy[faster-cache]==1.15.0",
    "types-urllib3>=1.26.25",
    "asyncpg-stubs>=0.29.0,<1.0.0",
    "sqlalchemy-stubs>=0.4.0,<1.0.0",
    "types-requests>=2.32.0",
    "types-boto3>=1.26.0,<2.0.0",
    "pre-commit==4.2.0"
]

[tool.coverage.run]
source = ["backend/app"]

[tool.ruff]
line-length = 120
target-version = "py311"
preview = true

[tool.ruff.lint]
select = [
    "F",  # pyflakes
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "I",  # isort
    "N",  # pep8-naming
    # "D101", "D103",  # pydocstyle missing docstring in public function/class
    "D201","D202","D207","D208","D214","D215","D300","D301","D417", "D419",  # pydocstyle
    "YTT",  # flake8-2020
    "ANN",  # flake8-annotations
    "ASYNC",  # flake8-async
    "S",  # flake8-bandit
    "BLE",  # flake8-blind-except
    "B",  # flake8-bugbear
    "A",  # flake8-builtins
    "COM",  # flake8-commas
    "CPY",  # flake8-copyright
    "C4",  # flake8-comprehensions
    "DTZ",  # flake8-datetimez
    "T10",  # flake8-debugger
    "ISC",  # flake8-implicit-str-concat
    "ICN",  # flake8-import-conventions
    "LOG",  # flake8-logging
    "PIE",  # flake8-pie
    "T20",  # flake8-print
    "PYI",  # flake8-pyi
    "PT",  # flake8-pytest-style
    "RET",  # flake8-return
    "SIM",  # flake8-simplify
    "ARG",  # flake8-unused-arguments
    "PTH",  # flake8-use-pathlib
    "PERF",  # perflint
    "NPY",  # numpy
    "FURB",  # refurb
    "RUF",  # ruff specific
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "B904",  # raise from
    "C901",  # too complex
    "F403",  # star imports
    "E731",  # lambda assignment
    "C416",  # list comprehension to list()
    "ANN002",  # missing type annotations on *args
    "ANN003",  # missing type annotations on **kwargs
    "COM812",  # trailing comma missing
    "ISC001",  # implicit string concatenation (handled by format)
]
exclude = [".git"]

[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"

[tool.ruff.lint.isort]
known-first-party = ["app", "tests", "fastemplate"]
known-third-party = ["fastapi"]

[tool.ruff.lint.per-file-ignores]
"**/__init__.py" = ["I001", "F401", "CPY001"]
"**/version.py" = ["CPY001"]
"scripts/**.py" = ["D", "T201", "S101", "ANN", "RUF030"]
".github/**.py" = ["D", "T201", "ANN"]
"backend/tests/**.py" = ["D103", "CPY001", "S101", "T201", "ANN001", "ANN201", "ANN202", "ARG001", "RUF029", "RUF030", "DTZ003"]
"backend/migrations/**.py" = ["ANN", "CPY001"]
"backend/app/main.py" = ["ANN"]
"backend/app/schemas/**.py" = ["A"]
"backend/app/models.py" = ["A"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
mypy_path = "backend/"
files = "backend/app"
show_error_codes = true
pretty = true
warn_unused_ignores = true
warn_redundant_casts = true
no_implicit_optional = true
check_untyped_defs = true
implicit_reexport = false
explicit_package_bases = true
plugins = ['pydantic.mypy']

[[tool.mypy.overrides]]
module = ["posthog", "magic", "boto3"]
ignore_missing_imports = true

[tool.ty.environment]
python-version = "3.11"
python-platform = "linux"

[tool.ty.rules]
unresolved-import = "warn"